move mailgun message sending functionality to messaging utils
